backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.2.1
pydantic>=2.5.0
openai>=1.0.0
python-dotenv>=1.0.0
pdfminer.six>=20231228
numpy>=1.24.0
scikit-learn>=1.3.0